Split test_transform_quantiles
